1.出库单增加导入，导入模板下载接口。

+++ business-center/wms-center/wms-service-api/wms-business-delivery-api/src/main/java/com/szmsd/delivery/dto/DelOutboundDetailImportDto2.java
@@ -21,5 +21,5 @@ public class DelOutboundDetailImportDto2 implements Serializable {
     private String sku;
 
     @ApiModelProperty(value = "数量")
-    private Long qty;
+    private Integer qty;
 }
